Python: Remove shared changes.
I had forgotten that `DataFlowImplCommon.qll` is shared across
languages, and so any change would have to be mirrored there.

But that would also mean that other languages would have to add the same
setup as us for these nodes, which I think is going to be a hard sell.

I'm not really sure where this leaves this proposal... :/