Python: Add FileSystemWriteAccess concept
I made `FileSystemWriteAccess` be a subclass of `FileSystemAccess` (like in [JS](https://github.com/github/codeql/blob/64001cc02cd67cc5df003199954be0d4135b3eb6/javascript/ql/src/semmle/javascript/Concepts.qll#L68-L74)), but then I started wondering about how I could  give a good result for `getAPathArgument`, and what would a good result even be? The argument to the `open` call, or the object that the `write` method is called on? I can't see how doing either of these enables us to do anything useful...

So I looked closer at how JS uses `FileSystemWriteAccess`:

1. as sink for zip-slip: https://github.com/github/codeql/blob/7c51dff0f781a0ca1efe32ef449b7b1b88cf0548/javascript/ql/src/semmle/javascript/security/dataflow/ZipSlipCustomizations.qll#L121
2. as sink for downloading unsafe files (identified through their extension) through non-secure connections: https://github.com/github/codeql/blob/89ef6ea4eb1d7f606380fc9cd904ac2b540c8eae/javascript/ql/src/semmle/javascript/security/dataflow/InsecureDownloadCustomizations.qll#L134-L150
3. as sink for writing untrusted data to a local file  https://github.com/github/codeql/blob/93b1e59d62a596fe40ebce13ba2dd8e6117b25fd/javascript/ql/src/semmle/javascript/security/dataflow/HttpToFileAccessCustomizations.qll#L43-L46

for the 2 first sinks, it's important that `getAPathArgument` has a proper result... so that solves the problem, and highlights that it _can_ be important to give proper results for `getAPathArgument` (if possible).

So I'm trying to do best effort for `f = open(...); f.write(...)`, but with this current code we won't always be able to give a result (as highlighted by the tests). It will also be the case that there are multiple `FileSystemAccess` with the same path-argument, which could be a little strange.

overall, I'm not super confident about the way this new concept and implementation turned out, but it also seems like the best I could come up with right now...

The obvious alternative solution is to NOT make `FileSystemWriteAccess` a subclass of `FileSystemAccess`, but I'm not very tempted to go down this path, given the examples of this being useful above, and just the general notion that we should be able to model writes as being a specialized kind of `FileSystemAccess`.