Java : Improve the existing SQL Injection query
The query in it's current form fails to recognise flows where a
tainted object is assigned to a class parameter. For example,
consider the following taint flow.

```
query.where = `taint`;

...

generateQuery(query){
	q = "query";
	q += query.where;
	return q;
}
```

This causes the query to have multiple false negatives such as in the case of
`OpenBravo-3.2.0` as described in [this blog](https://blog.sonarsource.com/exploiting-hibernate-injections)