Ruby: rename codeql-extractor.yml template
This file is copied into the extractor pack when it is built, but having
it in the source tree confuses the codeql CLI into thinking that
`codeql/ruby` is itself an extractor pack. This can cause errors when
running `codeql database create`.

This change renames the file to `codeql-extractor.template.yml`, so the
CLI ignores it.